+++ src/controllers/spellcheck/SpellChecker.cpp
@@ -12,6 +12,10 @@
 #include "util/FilesystemHelpers.hpp"
 #include "util/XDGDirectory.hpp"
 
+#include <algorithm>
+#include <QDir>
+#include <QLocale>
+
 #ifdef CHATTERINO_WITH_SPELLCHECK
 #    include <hunspell/hunspell.hxx>
 #endif
@@ -27,6 +31,11 @@ std::vector<DictionaryInfo> loadDictionariesFromDirectory(
     const QDir &searchDirectory, bool isSystem)
 {
     std::vector<DictionaryInfo> dictionaries;
+
+    if (!searchDirectory.exists())
+    {
+        return dictionaries;
+    }
 
     for (const auto &affInfo : searchDirectory.entryInfoList(
              {"*.aff"}, QDir::Files | QDir::NoDotAndDotDot, QDir::Name))
@@ -73,6 +82,164 @@ QString resolveDictionaryPath(const QString &path)
         return path;
     }
     return combinePath(getApp()->getPaths().dictionariesDirectory, path);
+}
+
+QString dictionaryLocaleName(const QString &baseName)
+{
+    const auto parts = baseName.split('_');
+    if (parts.size() >= 2)
+    {
+        return parts[0] + '_' + parts[1];
+    }
+    return baseName;
+}
+
+QString dictionaryDisplayName(const QString &baseName)
+{
+    const auto localeName = dictionaryLocaleName(baseName);
+    const QLocale locale(localeName);
+    const auto language = locale.nativeLanguageName();
+    if (language.isEmpty())
+    {
+        return baseName;
+    }
+
+    const auto territory = locale.nativeTerritoryName();
+    if (!territory.isEmpty())
+    {
+        return language + " (" + territory + ')';
+    }
+
+    return language;
+}
+
+void addLibreOfficeExtensionDirectories(
+    std::vector<std::pair<QString, bool>> &searchDirectories,
+    const QString &extensionsPath)
+{
+    QDir extensionsDir(extensionsPath);
+    if (!extensionsDir.exists())
+    {
+        return;
+    }
+
+    for (const auto &entry :
+         extensionsDir.entryList({"dict-*"}, QDir::Dirs | QDir::NoDotAndDotDot))
+    {
+        searchDirectories.emplace_back(
+            extensionsDir.absoluteFilePath(entry), true);
+    }
+}
+
+void addLibreOfficeInstallDirectories(
+    std::vector<std::pair<QString, bool>> &searchDirectories,
+    const QString &installRoot)
+{
+    if (installRoot.isEmpty())
+    {
+        return;
+    }
+
+    addLibreOfficeExtensionDirectories(
+        searchDirectories, combinePath(installRoot, "share/extensions"));
+
+    QDir installDir(installRoot);
+    if (!installDir.exists())
+    {
+        return;
+    }
+
+    for (const auto &entry :
+         installDir.entryList({"LibreOffice*"}, QDir::Dirs | QDir::NoDotAndDotDot))
+    {
+        addLibreOfficeExtensionDirectories(
+            searchDirectories,
+            combinePath(installDir.absoluteFilePath(entry), "share/extensions"));
+    }
+}
+
+void addWindowsDictionarySearchDirectories(
+    std::vector<std::pair<QString, bool>> &searchDirectories)
+{
+    QStringList installRoots;
+    for (const auto *envVar :
+         {"ProgramFiles", "ProgramFiles(x86)", "LocalAppData"})
+    {
+        const auto value = qEnvironmentVariable(envVar);
+        if (!value.isEmpty())
+        {
+            installRoots << value;
+        }
+    }
+
+    for (const auto &root : installRoots)
+    {
+        addLibreOfficeInstallDirectories(searchDirectories, root + "/LibreOffice");
+        addLibreOfficeInstallDirectories(searchDirectories, root + "/Programs/LibreOffice");
+
+        searchDirectories.emplace_back(root + "/Hunspell/Dictionaries", true);
+        searchDirectories.emplace_back(root + "/Hunspell/dic", true);
+        searchDirectories.emplace_back(root + "/OpenOffice 4/share/dict", true);
+        searchDirectories.emplace_back(root + "/OpenOffice/share/dict", true);
+    }
+}
+
+void addMacDictionarySearchDirectories(
+    std::vector<std::pair<QString, bool>> &searchDirectories)
+{
+    const QStringList hunspellDirectories = {
+        "/opt/homebrew/share/hunspell",
+        "/usr/local/share/hunspell",
+        "/opt/local/share/hunspell",
+        "/Library/Spelling",
+    };
+
+    for (const auto &directory : hunspellDirectories)
+    {
+        searchDirectories.emplace_back(directory, true);
+    }
+
+    addLibreOfficeExtensionDirectories(
+        searchDirectories,
+        "/Applications/LibreOffice.app/Contents/Resources/extensions");
+}
+
+std::vector<std::pair<QString, bool>> getDictionarySearchDirectories()
+{
+    std::vector<std::pair<QString, bool>> searchDirectories{
+        {getApp()->getPaths().dictionariesDirectory, false},
+    };
+
+#    if defined(Q_OS_UNIX) and !defined(Q_OS_DARWIN)
+    auto dataDirs = getXDGBaseDirectories(XDGDirectoryType::Data);
+    for (const auto &dataDir : dataDirs)
+    {
+        searchDirectories.emplace_back(combinePath(dataDir, "hunspell"), true);
+        searchDirectories.emplace_back(combinePath(dataDir, "myspell"), true);
+        searchDirectories.emplace_back(combinePath(dataDir, "myspell/dicts"),
+                                       true);
+    }
+#    elif defined(Q_OS_WIN)
+    addWindowsDictionarySearchDirectories(searchDirectories);
+#    elif defined(Q_OS_DARWIN)
+    addMacDictionarySearchDirectories(searchDirectories);
+#    endif
+
+    return searchDirectories;
+}
+
+QString dictionaryListLabel(const DictionaryInfo &dict)
+{
+    auto label = dictionaryDisplayName(dict.name);
+    if (label != dict.name)
+    {
+        label += " [" + dict.name + ']';
+    }
+    if (dict.isSystem)
+    {
+        label += " (System)";
+    }
+    return label;
 }
 
 }  // namespace
@@ -137,9 +304,8 @@ SpellCheckerPrivate::SpellCheckerPrivate(const char *affpath, const char *dpath)
 }
 
 SpellChecker::SpellChecker()
-    : private_(SpellCheckerPrivate::tryLoad(
-          getSettings()->spellCheckingDefaultDictionary))
 {
+    this->loadDictionary(getSettings()->spellCheckingDefaultDictionary);
 }
 #else
 class SpellCheckerPrivate
@@ -149,6 +315,23 @@ SpellChecker::SpellChecker() = default;
 #endif
 
 SpellChecker::~SpellChecker() = default;
+
+bool SpellChecker::loadDictionary(const QString &path)
+{
+#ifdef CHATTERINO_WITH_SPELLCHECK
+    this->private_ = SpellCheckerPrivate::tryLoad(path);
+    this->loadedDictionaryPath_ = this->private_ ? path : QString{};
+    return this->private_ != nullptr || path.isEmpty();
+#else
+    (void)path;
+    return false;
+#endif
+}
+
+const QString &SpellChecker::loadedDictionaryPath() const
+{
+    return this->loadedDictionaryPath_;
+}
 
 bool SpellChecker::isLoaded() const
 {
@@ -197,55 +380,24 @@ std::vector<std::string> SpellChecker::suggestions(const QString &word)
 std::vector<DictionaryInfo> SpellChecker::getAvailableDictionaries() const
 {
 #ifdef CHATTERINO_WITH_SPELLCHECK
-    std::vector<std::pair<QString, bool>> searchDirectories{
-        {getApp()->getPaths().dictionariesDirectory, false},
-    };
-
-#    if defined(Q_OS_UNIX) and !defined(Q_OS_DARWIN)
-    // For each XDG data directory, search in hunspell, myspell, and myspell/dicts.
-    // This somewhat matches where dictionaries are stored on Ubuntu & Fedora
-    // if we want to support defaulting to your LC_ALL language.
-
-    auto dataDirs = getXDGBaseDirectories(XDGDirectoryType::Data);
-    for (const auto &dataDir : dataDirs)
-    {
-        searchDirectories.emplace_back(combinePath(dataDir, "hunspell"), true);
-        searchDirectories.emplace_back(combinePath(dataDir, "myspell"), true);
-        searchDirectories.emplace_back(combinePath(dataDir, "myspell/dicts"),
-                                       true);
-    }
-#    endif
-
     std::vector<DictionaryInfo> dictionaries;
 
-    for (const auto &[searchDirectory, isSystem] : searchDirectories)
+    for (const auto &[searchDirectory, isSystem] :
+         getDictionarySearchDirectories())
     {
         qCDebug(chatterinoSpellcheck)
             << "Looking for dictionaries in" << searchDirectory
             << "isSystem:" << isSystem;
         for (const auto &dict :
-             loadDictionariesFromDirectory(searchDirectory, isSystem))
+             loadDictionariesFromDirectory(QDir(searchDirectory), isSystem))
         {
             if (dict.isSymbolicLink && dict.isSystem)
             {
-                // NOTE: We currently filter out symbolic links from system-loaded dictionaries.
-                // Without this, the list of dictionaries we "support" would be too high on Linux distros.
-                // As an example, this is the symlinks the installation of `hunspell-en-gb` creates on Arch Linux:
-                // - en_AG.aff -> en_GB-large.aff
-                // -  ...
-                // - en_GB-large.aff
                 continue;
             }
-            auto name = [&] -> QString {
-                if (dict.isSystem)
-                {
-                    return dict.name % " (System)";
-                }
-                return dict.name;
-            }();
 
             dictionaries.push_back(DictionaryInfo{
-                .name = name,
+                .name = dictionaryListLabel(dict),
                 .path = dict.path,
                 .isSymbolicLink = dict.isSymbolicLink,
                 .isSystem = dict.isSystem,
@@ -258,6 +410,11 @@ std::vector<DictionaryInfo> SpellChecker::getAvailableDictionaries() const
                           return std::tie(lhs.isSystem, lhs.name, lhs.path) <
                                  std::tie(rhs.isSystem, rhs.name, rhs.path);
                       });
+
+    const auto uniqueEnd = std::ranges::unique(
+                               dictionaries, {}, &DictionaryInfo::path)
+                               .begin();
+    dictionaries.erase(uniqueEnd, dictionaries.end());
 
     return dictionaries;
 #else